Estimate known outcomes from case and death time-series data

Calculates how the severity of a disease changes over time, corrected for a user-specified delay. If cases are supplied, and the delay distribution representing the delay between case detection and death, then a case fatality ratio over time is estimated

Usage

estimate_time_varying(
  df_in,
  epi_dist,
  burn_in_value = get_default_burn_in(epi_dist),
  smooth_inputs = FALSE,
  smoothing_window = 7,
  correct_for_delays = TRUE
)

Arguments

df_in

A data.frame containing the outbreak data. A daily time series with dates or some other absolute indicator of time (e.g. epiday/epiweek) and the numbers of new cases and new deaths at each time point. Note that the required columns are "date" (for the date), "cases" (for the number of reported cases), and "deaths" (for the number of reported deaths) on each day of the outbreak. Note that the data.frame is required to have an unbroken sequence of dates with no missing dates in between. Note also that the total number of cases must be greater than the total number of reported deaths.

epi_dist

The delay distribution used, in the form of an epiparameter::epidist() object. This is used to obtain a probability mass function parameterised by time; i.e. \(f(t)\) which gives the probability a case has a known outcomes (i.e. death) at time \(t\), parameterised with disease-specific parameters before it is supplied here. A typical example would be a symptom onset to death delay distribution.

burn_in_value

The number of time-points (typically days) to disregard at the start of the time-series, if a burn-in period is desired. The default value is set to the mean of the central spread of the epidist object passed to the function, assuming the temporal resolution is daily. Alternatively, a sensible value might be 7, to disregard the first week of cases and deaths. To consider all case data including the start of the time-series, set this argument to 0.

smooth_inputs

A boolean flag determining whether the user wishes to smooth the case and death time-series, using a moving average procedure before calculating the time-varying severity. Useful for noisey time-series or time-series with strong reporting (e.g., weekend) effects

smoothing_window

A number determining the smoothing window size to use when smoothing the case and death time-series, using a moving average procedure before calculating the time-varying severity. Useful for noisy time-series or time-series with strong reporting (e.g., weekend) effects. The default value is 1 for no smoothing. Values > 1 apply smoothing.

correct_for_delays

A boolean flag indicating whether the user wishes to correct for the delay between case detection and death. FALSE corresponds to a naive severity being calculated, TRUE corresponds to the user calculating a corrected severity

Value

A data.frame containing the MLE estimate and 95% confidence interval of the corrected severity

Examples

library(datadelay)
library(epiparameter)
library(covidregionaldata)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

df_covid_uk <- get_national_data(
  countries = "united kingdom", source = "who", verbose = FALSE
)
df_covid_uk <- dplyr::rename(
  df_covid_uk,
  cases = cases_new, deaths = deaths_new
)

df_covid_uk_subset <- subset(df_covid_uk, date <= "2020-12-31")

onset_to_death_covid <- epidist_db(
  disease = "COVID-19",
  epi_dist = "onset_to_death",
  author = "Linton_etal"
)
#> Warning: More than one study found. Selecting first one. Please report an issue with duplicated studies.
#> Using Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150. 
#> To retrieve the short citation use the 'get_citation' function

df_covid_cfr_uk_naive <- estimate_time_varying(
  df_covid_uk_subset,
  epi_dist = onset_to_death_covid,
  smooth_inputs = TRUE,
  burn_in_value = 7L,
  correct_for_delays = FALSE
)

df_covid_cfr_uk_corrected <- estimate_time_varying(
  df_covid_uk_subset,
  epi_dist = onset_to_death_covid,
  smooth_inputs = TRUE,
  burn_in_value = 7L,
  correct_for_delays = TRUE
)