Estimate known outcomes from case and death time-series data

Calculates how the severity of a disease changes over time, corrected for a user-specified delay. If cases are supplied, and the delay distribution representing the delay between case detection and death, then a case fatality ratio over time is estimated

Usage

estimate_time_varying(
  data,
  epi_dist = NULL,
  burn_in_value = get_default_burn_in(epi_dist),
  smooth_inputs = FALSE,
  smoothing_window = 1,
  correct_for_delays = TRUE
)

Arguments

data

A data.frame containing the outbreak data. A daily time series with dates or some other absolute indicator of time (e.g. epiday/epiweek) and the numbers of new cases and new deaths at each time point. Note that the required columns are "date" (for the date), "cases" (for the number of reported cases), and "deaths" (for the number of reported deaths) on each day of the outbreak. Note that the data.frame is required to have an unbroken sequence of dates with no missing dates in between. The "date" column must be of the class Date (see as.Date()). Note also that the total number of cases must be greater than the total number of reported deaths.

epi_dist

The delay distribution used, in the form of an epiparameter::epidist() object. This is used to obtain a probability mass function parameterised by time; i.e. \(f(t)\) which gives the probability a case has a known outcomes (i.e. death) at time \(t\), parameterised with disease-specific parameters before it is supplied here. A typical example would be a symptom onset to death delay distribution.

burn_in_value

The number of time-points (typically days) to disregard at the start of the time-series, if a burn-in period is desired. The default value is set to the mean of the central spread of the <epidist> object passed to the function, assuming the temporal resolution is daily. Alternatively, a sensible value might be 7, to disregard the first week of cases and deaths. To consider all case data including the start of the time-series, set this argument to 1.

smooth_inputs

A boolean flag determining whether the user wishes to smooth the case and death time-series, using a moving average procedure before calculating the time-varying severity. Useful for noisy time-series or time-series with strong reporting (e.g., weekend) effects

smoothing_window

An odd number determining the smoothing window size to use when smoothing the case and death time-series, using a moving average procedure before calculating the time-varying severity. Useful for noisy time-series or time-series with strong reporting (e.g., weekend) effects. The default value is 1 for no smoothing. Values > 1 apply smoothing.

correct_for_delays

A boolean flag indicating whether the user wishes to correct for the delay between case detection and death. FALSE corresponds to a naive severity being calculated, TRUE corresponds to the user calculating a corrected severity

Value

A data.frame containing the MLE estimate and 95% confidence interval of the corrected severity

Examples

# load package data
data("ebola1976")

# get an onset to death distribution from the {epiparameter} package
onset_to_death_ebola <- epiparameter::epidist_db(
  disease = "Ebola Virus Disease",
  epi_dist = "onset_to_death",
  author = "Barry_etal"
)
#> Using Barry et al. (2018) <10.1016/S0140-6736(18)31387-4> PMID: 30047375. 
#> To retrieve the short citation use the 'get_citation' function

# estimate time varying severity without correcting for delays
estimate_time_varying(
  data = ebola1976,
  smooth_inputs = TRUE,
  burn_in_value = 7L,
  correct_for_delays = FALSE
)
#>          date cases deaths severity_me severity_lo severity_hi
#> 1  1976-08-25     1      0          NA          NA          NA
#> 2  1976-08-26     0      0          NA          NA          NA
#> 3  1976-08-27     0      0          NA          NA          NA
#> 4  1976-08-28     0      0          NA          NA          NA
#> 5  1976-08-29     0      0          NA          NA          NA
#> 6  1976-08-30     0      0          NA          NA          NA
#> 7  1976-08-31     0      0          NA          NA          NA
#> 8  1976-09-01     1      0   0.0000000 0.000000000   0.9750000
#> 9  1976-09-02     1      0   0.0000000 0.000000000   0.9750000
#> 10 1976-09-03     1      0   0.0000000 0.000000000   0.9750000
#> 11 1976-09-04     4      0   0.0000000 0.000000000   0.6023646
#> 12 1976-09-05     1      0   0.0000000 0.000000000   0.9750000
#> 13 1976-09-06     1      0   0.0000000 0.000000000   0.9750000
#> 14 1976-09-07     3      0   0.0000000 0.000000000   0.7075982
#> 15 1976-09-08     2      1   0.5000000 0.012579117   0.9874209
#> 16 1976-09-09     7      2   0.2857143 0.036692566   0.7095791
#> 17 1976-09-10     9      1   0.1111111 0.002809137   0.4824965
#> 18 1976-09-11     4      3   0.7500000 0.194120450   0.9936905
#> 19 1976-09-12     8      1   0.1250000 0.003159724   0.5265097
#> 20 1976-09-13     9      2   0.2222222 0.028144973   0.6000936
#> 21 1976-09-14     8      1   0.1250000 0.003159724   0.5265097
#> 22 1976-09-15     9      8   0.8888889 0.517503485   0.9971909
#> 23 1976-09-16     6      4   0.6666667 0.222778096   0.9567281
#> 24 1976-09-17     6      3   0.5000000 0.118117249   0.8818828
#> 25 1976-09-18     3      5          NA          NA          NA
#> 26 1976-09-19     8      8   1.0000000 0.630583352   1.0000000
#> 27 1976-09-20    10      6   0.6000000 0.262378077   0.8784477
#> 28 1976-09-21     5      6          NA          NA          NA
#> 29 1976-09-22    16     10   0.6250000 0.354346094   0.8480163
#> 30 1976-09-23    12      3   0.2500000 0.054860645   0.5718585
#> 31 1976-09-24    11      7   0.6363636 0.307904715   0.8907366
#> 32 1976-09-25    11      8   0.7272727 0.390257440   0.9397823
#> 33 1976-09-26     7     11          NA          NA          NA
#> 34 1976-09-27     7      7   1.0000000 0.590383603   1.0000000
#> 35 1976-09-28     4     13          NA          NA          NA
#> 36 1976-09-29     4     12          NA          NA          NA
#> 37 1976-09-30     3      9          NA          NA          NA
#> 38 1976-10-01     7      9          NA          NA          NA
#> 39 1976-10-02     2      7          NA          NA          NA
#> 40 1976-10-03     3      6          NA          NA          NA
#> 41 1976-10-04     8      8   1.0000000 0.630583352   1.0000000
#> 42 1976-10-05     5      6          NA          NA          NA
#> 43 1976-10-06     7      2   0.2857143 0.036692566   0.7095791
#> 44 1976-10-07     5      6          NA          NA          NA
#> 45 1976-10-08     2      4          NA          NA          NA
#> 46 1976-10-09     3      3   1.0000000 0.292401774   1.0000000
#> 47 1976-10-10     4      2   0.5000000 0.067585986   0.9324140
#> 48 1976-10-11     3      6          NA          NA          NA
#> 49 1976-10-12     6      4   0.6666667 0.222778096   0.9567281
#> 50 1976-10-13     2      6          NA          NA          NA
#> 51 1976-10-14     1      7          NA          NA          NA
#> 52 1976-10-15     0      7          NA          NA          NA
#> 53 1976-10-16     0      2          NA          NA          NA
#> 54 1976-10-17     0      4          NA          NA          NA
#> 55 1976-10-18     1      1   1.0000000 0.025000000   1.0000000
#> 56 1976-10-19     1      1   1.0000000 0.025000000   1.0000000
#> 57 1976-10-20     2      4          NA          NA          NA
#> 58 1976-10-21     0      1          NA          NA          NA
#> 59 1976-10-22     0      1          NA          NA          NA
#> 60 1976-10-23     0      0          NA          NA          NA
#> 61 1976-10-24     1      0   0.0000000 0.000000000   0.9750000
#> 62 1976-10-25     0      0          NA          NA          NA
#> 63 1976-10-26     0      2          NA          NA          NA
#> 64 1976-10-27     0      0          NA          NA          NA
#> 65 1976-10-28     0      2          NA          NA          NA
#> 66 1976-10-29     0      1          NA          NA          NA
#> 67 1976-10-30     0      0          NA          NA          NA
#> 68 1976-10-31     0      0          NA          NA          NA
#> 69 1976-11-01     0      0          NA          NA          NA
#> 70 1976-11-02     0      0          NA          NA          NA
#> 71 1976-11-03     0      0          NA          NA          NA
#> 72 1976-11-04     0      0          NA          NA          NA
#> 73 1976-11-05     0      1          NA          NA          NA

# estimate time varying severity while correcting for delays
estimate_time_varying(
  data = ebola1976,
  epi_dist = onset_to_death_ebola,
  smooth_inputs = TRUE,
  burn_in_value = 7L,
  correct_for_delays = TRUE
)
#>          date cases deaths severity_me severity_lo severity_hi
#> 1  1976-08-25     1      0          NA          NA          NA
#> 2  1976-08-26     0      0          NA          NA          NA
#> 3  1976-08-27     0      0          NA          NA          NA
#> 4  1976-08-28     0      0          NA          NA          NA
#> 5  1976-08-29     0      0          NA          NA          NA
#> 6  1976-08-30     0      0          NA          NA          NA
#> 7  1976-08-31     0      0          NA          NA          NA
#> 8  1976-09-01     1      0          NA          NA          NA
#> 9  1976-09-02     1      0          NA          NA          NA
#> 10 1976-09-03     1      0          NA          NA          NA
#> 11 1976-09-04     4      0          NA          NA          NA
#> 12 1976-09-05     1      0          NA          NA          NA
#> 13 1976-09-06     1      0          NA          NA          NA
#> 14 1976-09-07     3      0          NA          NA          NA
#> 15 1976-09-08     2      1          NA          NA          NA
#> 16 1976-09-09     7      2          NA          NA          NA
#> 17 1976-09-10     9      1          NA          NA          NA
#> 18 1976-09-11     4      3          NA          NA          NA
#> 19 1976-09-12     8      1          NA          NA          NA
#> 20 1976-09-13     9      2          NA          NA          NA
#> 21 1976-09-14     8      1   1.0000000 0.025000000   1.0000000
#> 22 1976-09-15     9      8          NA          NA          NA
#> 23 1976-09-16     6      4          NA          NA          NA
#> 24 1976-09-17     6      3          NA          NA          NA
#> 25 1976-09-18     3      5          NA          NA          NA
#> 26 1976-09-19     8      8          NA          NA          NA
#> 27 1976-09-20    10      6          NA          NA          NA
#> 28 1976-09-21     5      6          NA          NA          NA
#> 29 1976-09-22    16     10          NA          NA          NA
#> 30 1976-09-23    12      3   0.6000000 0.146632800   0.9472550
#> 31 1976-09-24    11      7          NA          NA          NA
#> 32 1976-09-25    11      8          NA          NA          NA
#> 33 1976-09-26     7     11          NA          NA          NA
#> 34 1976-09-27     7      7          NA          NA          NA
#> 35 1976-09-28     4     13          NA          NA          NA
#> 36 1976-09-29     4     12          NA          NA          NA
#> 37 1976-09-30     3      9          NA          NA          NA
#> 38 1976-10-01     7      9          NA          NA          NA
#> 39 1976-10-02     2      7   0.8750000 0.473490329   0.9968403
#> 40 1976-10-03     3      6   0.7500000 0.349144206   0.9681460
#> 41 1976-10-04     8      8   0.8888889 0.517503485   0.9971909
#> 42 1976-10-05     5      6   0.6666667 0.299295056   0.9251454
#> 43 1976-10-06     7      2   0.2222222 0.028144973   0.6000936
#> 44 1976-10-07     5      6   0.7500000 0.349144206   0.9681460
#> 45 1976-10-08     2      4   0.5000000 0.157012770   0.8429872
#> 46 1976-10-09     3      3   0.3750000 0.085233414   0.7551368
#> 47 1976-10-10     4      2   0.2857143 0.036692566   0.7095791
#> 48 1976-10-11     3      6   0.8571429 0.421276803   0.9963897
#> 49 1976-10-12     6      4   0.6666667 0.222778096   0.9567281
#> 50 1976-10-13     2      6   1.0000000 0.540741874   1.0000000
#> 51 1976-10-14     1      7          NA          NA          NA
#> 52 1976-10-15     0      7          NA          NA          NA
#> 53 1976-10-16     0      2   0.3333333 0.043271868   0.7772219
#> 54 1976-10-17     0      4   0.8000000 0.283582064   0.9949492
#> 55 1976-10-18     1      1   0.2000000 0.005050763   0.7164179
#> 56 1976-10-19     1      1   0.2000000 0.005050763   0.7164179
#> 57 1976-10-20     2      4   0.8000000 0.283582064   0.9949492
#> 58 1976-10-21     0      1   0.2000000 0.005050763   0.7164179
#> 59 1976-10-22     0      1   0.2500000 0.006309463   0.8058796
#> 60 1976-10-23     0      0   0.0000000 0.000000000   0.6023646
#> 61 1976-10-24     1      0   0.0000000 0.000000000   0.6023646
#> 62 1976-10-25     0      0   0.0000000 0.000000000   0.7075982
#> 63 1976-10-26     0      2   0.6666667 0.094299324   0.9915962
#> 64 1976-10-27     0      0   0.0000000 0.000000000   0.8418861
#> 65 1976-10-28     0      2   1.0000000 0.158113883   1.0000000
#> 66 1976-10-29     0      1   0.5000000 0.012579117   0.9874209
#> 67 1976-10-30     0      0   0.0000000 0.000000000   0.8418861
#> 68 1976-10-31     0      0   0.0000000 0.000000000   0.9750000
#> 69 1976-11-01     0      0   0.0000000 0.000000000   0.9750000
#> 70 1976-11-02     0      0   0.0000000 0.000000000   0.9750000
#> 71 1976-11-03     0      0   0.0000000 0.000000000   0.9750000
#> 72 1976-11-04     0      0   0.0000000 0.000000000   0.9750000
#> 73 1976-11-05     0      1          NA          NA          NA